requirements.txt
Flask==2.0.3
tensorflow==2.8.0
Pillow==9.0.1
